# Change flag storage to reduce memory use